Moving browser-steps to its own Blueprint
